[Imp] PP20: Validate header more thoroughly. [Imp] UMX: Stop reading tables if EOF is reached. [Fix] J2B: Import non-existing patterns as empty patterns.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5887 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ITTools.h
@@ -100,7 +100,7 @@ struct PACKED ITEnvelope
 	// Convert OpenMPT's internal envelope format to an IT/MPTM envelope.
 	void ConvertToIT(const InstrumentEnvelope &mptEnv, uint8 envOffset, uint8 envDefault);
 	// Convert IT/MPTM envelope data into OpenMPT's internal envelope format - To be used by ITInstrToMPT()
-	void ConvertToMPT(InstrumentEnvelope &mptEnv, uint8 envOffset, int maxNodes) const;
+	void ConvertToMPT(InstrumentEnvelope &mptEnv, uint8 envOffset, uint8 maxNodes) const;
 };
 
 STATIC_ASSERT(sizeof(ITEnvelope) == 82);
